+++ plugins/GSdx/GSVertexList.h
@@ -33,7 +33,7 @@ public:
 	{
 		m_base = _aligned_malloc(sizeof(Vertex) * countof(m_v), 32);
 
-		for(unsigned int i = 0; i < countof(m_v); i++)
+		for(size_t i = 0; i < countof(m_v); i++)
 		{
 			m_v[i] = &((Vertex*)m_base)[i];
 		}
